django_backend/requirements.txt
# Django Core
Django>=4.2,<5.0
django-cors-headers>=4.3

# PostgreSQL Database
psycopg2-binary>=2.9
dj-database-url>=2.1

# REST Framework
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3

# Environment Variables
python-dotenv>=1.0

# Password Hashing
bcrypt>=4.1

# Image Processing (for inline images)
Pillow>=10.0

# API Documentation (optional)
drf-yasg>=1.21

# Utilities
pytz>=2024.1

# Production server
gunicorn>=21.2
whitenoise>=6.6
# SMS Integration
africastalking>=1.2.7